FM-15

+++ database/migrations/2025_08_28_125901_created_produto_tamanho_table.php
@@ -11,19 +11,20 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('produto_tamanho', function (Blueprint $table) {
-            $table->foreignId('produto_id')->constrained()->onDelete('cascade');
-            $table->foreignId('tamanho_id')->constrained()->onDelete('cascade');
-            $table->decimal('preco', 8, 2);
-            $table->primary(['produto_id', 'tamanho_id']);
-        });
+        if (!Schema::hasTable('produto_tamanho')) {
+            Schema::create('produto_tamanho', function (Blueprint $table) {
+                $table->foreignId('produto_id')->constrained()->onDelete('cascade');
+                $table->foreignId('tamanho_id')->constrained()->onDelete('cascade');
+                $table->decimal('preco', 8, 2);
+                $table->primary(['produto_id', 'tamanho_id']);
+            });
+        }
     }
-
     /**
      * Reverse the migrations.
      */
     public function down(): void
     {
-        //
+        Schema::dropIfExists('produto_tamanho');
     }
 };
